Load filter bias from BRAM in init cycles.
Registers are resetted at the beginning of every action to the DFE.
Bias are saved in LUTs throughout the inputs, but not persisted over
actions. Hence, we store it in Block Ram (minimally, with only one
port read / write a weight value). At the beginning of every action,
the bias are load from bram in num_outputs clock cycles.

This is currently implemented only on CPU-initialized weights.